requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
astroid==2.9.3
async-timeout==4.0.2
attrs==21.4.0
boto3==1.20.49
botocore==1.23.49
cachetools==5.0.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==7.1.2
cloud-sql-python-connector==0.5.2
cryptography==36.0.2
Flask==2.0.3
frozenlist==1.3.0
google-api-core==2.7.1
google-api-python-client==2.41.0
google-auth==2.6.2
google-auth-httplib2==0.1.0
googleapis-common-protos==1.56.0
greenlet==1.1.2
httplib2==0.20.4
idna==3.3
iniconfig==1.1.1
isort==5.10.1
itsdangerous==2.1.0
Jinja2==3.0.3
jmespath==0.10.0
lazy-object-proxy==1.7.1
MarkupSafe==2.1.0
mccabe==0.6.1
multidict==6.0.2
packaging==21.3
platformdirs==2.4.1
pluggy==1.0.0
protobuf==3.19.4
py==1.11.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pylint==2.12.2
PyMySQL==1.0.2
pyOpenSSL==22.0.0
pyparsing==3.0.7
pytest==7.0.0
python-dateutil==2.8.2
python-dotenv==0.19.2
PyYAML==5.1
requests==2.27.1
rsa==4.8
s3transfer==0.5.1
six==1.16.0
slack-bolt==1.11.4
slack-sdk==3.13.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
toml==0.10.2
tomli==2.0.0
typing_extensions==4.0.1
uritemplate==4.1.1
urllib3==1.26.8
Werkzeug==2.0.3
wrapt==1.13.3
yarl==1.7.2
